# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "法律智能问答系统 API",
        "title": "LexVeritas API",
        "contact": {
            "name": "LexVeritas nsxzhou",
            "email": "1790146932@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/me": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新当前用户的姓名、手机、头像等信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新个人资料",
                "parameters": [
                    {
                        "description": "更新请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/users/me/password": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "修改当前用户密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "密码修改成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "旧密码错误或请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/users/me/quota": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "查看当前用户的Token配额使用情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取当前用户Token配额",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_dto.UsageStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_lexveritas_lex-veritas-backend_internal_dto.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "newPassword",
                "oldPassword"
            ],
            "properties": {
                "newPassword": {
                    "type": "string",
                    "minLength": 8
                },
                "oldPassword": {
                    "type": "string"
                }
            }
        },
...
        "github_com_lexveritas_lex-veritas-backend_internal_dto.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "github_com_lexveritas_lex-veritas-backend_internal_dto.UsageStats": {
            "type": "object",
            "properties": {
                "remaining": {
                    "type": "integer"
                },
                "tokenQuota": {
                    "type": "integer"
                },
                "tokenUsed": {
                    "type": "integer"
                },
                "usageRate": {
                    "description": "0-1",
                    "type": "number"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_lexveritas_lex-veritas-backend_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "输入格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}